Integrate lighting init control
